# swagger.yaml
swagger: "2.0"
info:
  description: |
    An API used to fetch information about dataset recipes.
  version: "1.0.0"
  title: "ONS Recipe API"
  license:
    name: "Open Government Licence v3.0"
    url: "http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/"
schemes:
- "http"
parameters:
  recipe_id:
    name: recipe_id
    description: "A unique id for a recipe"
    type: string
    in: path
    required: true
paths:
  /recipes:
    get:
      summary: "Get a list of recipes"
      description: |
        Returns a paged list of recipes
      produces:
      - "application/json"
      responses:
        200:
          description: "A list of recipes"
          schema:
            $ref: '#/definitions/RecipeList'
        500:
          $ref: '#/responses/InternalError'
  /recipes/{recipe_id}:
    get:
      summary: "Get a specific recipe"
      description: |
        Returns a specific recipe
      parameters:
      - $ref: '#/parameters/recipe_id'
      produces:
      - "application/json"
      responses:
        200:
          description: "A specific recipe"
          schema:
            $ref: '#/definitions/Recipe'
        404:
          description: "Recipe not found"
        500:
          $ref: '#/responses/InternalError'
responses:
  InternalError:
    description: "Failed to process the request due to an internal error"
definitions:
  RecipeList:
    type: object
    description: "A list of recipes"
    properties:
      count:
        type: number
        readOnly: true
        description: "The number of recipes returned"
      total_count:
        type: number
        readOnly: true
        description: "The total number of recipes"
      start:
        type: number
        readOnly: true
        description: "The start index for the results"
      items_per_page:
        type: number
        readOnly: true
        description: "The number of items per page"
      items:
        type: array
        readOnly: true
        items:
          $ref: '#/definitions/Recipe'
  Recipe:
    type: object
    description: "An object returned for a recipe"
    properties:
      id:
        type: string
        readOnly: true
        description: "A unique id for the recipe"
      alias:
        type: string
        readOnly: true
        description: "An alias for the recipe"
      format:
        type: string
        readOnly: true
        description: "The recipe format, e.g. v4"
      input_files:
        type: array
        items: 
          $ref: '#/definitions/InputFile'
      output_instances:
        type: number
        readOnly: true
        description: "The number of output instances"
  InputFile:
    type: object
    description: "An input file"
    properties:
      description: 
        type: string
        readOnly: true
        description: "The description of the input file"